backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
supabase==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
apscheduler==3.10.4
numpy==1.26.4
python-multipart==0.0.12
